use exsiting pm controller test file for pm integration tests w/ burnettk
