--- spiffworkflow-backend/tests/spiffworkflow_backend/unit/test_process_model_metadata_backfill_integration.py
@@ -1,79 +1,0 @@
-from unittest.mock import MagicMock
-from unittest.mock import patch
-
-from flask import Flask
-from starlette.testclient import TestClient
-
-from spiffworkflow_backend.routes.process_api_blueprint import _get_process_model
-from tests.spiffworkflow_backend.helpers.base_test import BaseTest
-
-
-class TestProcessModelMetadataBackfillIntegration(BaseTest):
-    @patch("spiffworkflow_backend.routes.process_models_controller.trigger_metadata_backfill")
-    def test_process_model_update_triggers_metadata_backfill(
-        self,
-        mock_trigger_metadata_backfill: MagicMock,
-        app: Flask,
-        client: TestClient,
-        with_db_and_bpmn_file_cleanup: None,
-    ) -> None:
-        user = self.create_user_with_permission("test_user")
-        process_model = self.create_process_model_with_metadata()
-        mock_trigger_metadata_backfill.return_value = {"status": "triggered", "task_id": "test-task-id"}
-
-        modified_process_model_identifier = process_model.id.replace("/", ":")
-        new_metadata_paths = [
-            {"key": "awesome_var", "path": "outer.inner"},
-            {"key": "invoice_number", "path": "invoice_number"},
-            {"key": "new_key", "path": "new.path"},
-        ]
-
-        response = client.put(
-            f"/v1.0/process-models/{modified_process_model_identifier}",
-            json={
-                "metadata_extraction_paths": new_metadata_paths,
-                "display_name": process_model.display_name,
-                "description": process_model.description,
-            },
-            headers=self.logged_in_headers(user),
-        )
-        assert response.status_code == 200
-
-        mock_trigger_metadata_backfill.assert_called_once()
-        args, _ = mock_trigger_metadata_backfill.call_args
-        _process_model_identifier, old_metadata_extraction_paths, new_metadata_extraction_paths = args
-
-        assert len(old_metadata_extraction_paths) == 2
-        assert old_metadata_extraction_paths[0]["key"] == "awesome_var"
-        assert old_metadata_extraction_paths[1]["key"] == "invoice_number"
-
-        assert len(new_metadata_extraction_paths) == 3
-        assert new_metadata_extraction_paths[0]["key"] == "awesome_var"
-        assert new_metadata_extraction_paths[1]["key"] == "invoice_number"
-        assert new_metadata_extraction_paths[2]["key"] == "new_key"
-
-    @patch("spiffworkflow_backend.routes.process_models_controller.trigger_metadata_backfill")
-    def test_process_model_update_without_metadata_changes_does_not_trigger_backfill(
-        self,
-        mock_trigger_metadata_backfill: MagicMock,
-        app: Flask,
-        client: TestClient,
-        with_db_and_bpmn_file_cleanup: None,
-    ) -> None:
-        user = self.create_user_with_permission("test_user")
-        process_model = self.create_process_model_with_metadata()
-        modified_process_model_identifier = process_model.id.replace("/", ":")
-
-        response = client.put(
-            f"/v1.0/process-models/{modified_process_model_identifier}",
-            json={
-                "display_name": "Updated Display Name",
-                "description": process_model.description,
-            },
-            headers=self.logged_in_headers(user),
-        )
-        assert response.status_code == 200
-
-        mock_trigger_metadata_backfill.assert_not_called()
-        updated_model = _get_process_model(process_model.id)
-        assert updated_model.display_name == "Updated Display Name"
